GenerateQuizScreen now fetches tokens from API
Token is fetched from API to ensure user is not asked same question multiple times. Token is resettable via a button in the info modal. User is notified if their quiz generation failed because the API no longer has enough new questions to satisfy the query and user is asked if they would like to reset the token. Token is also saved in AsyncStorage to persist in memory through app reloads.