Dark mode improvements in GenerateQuizScreen, StatusBar and header (Drawer- and StackNavigator). Fully custom picker component with JSDoc documentation.
Added dark mode to header and StatusBar. Greatly improved the looks of GenerateQuizScreen and made a few tweaks, such as removing the manual reset button, opting to only pop up a reset button if the token actually runs out of questions. Also removed user switch to opt in and out of using a token and token is now always used. Added a completely custom picker element CustomPicker, which has neat features and a decent JSDoc documentation to help other team members understand how it is used (you can see this documentation by hovering over the component in VS Code after importing the component and placing it in code)