Added a custom stepper component to make choosing amount of questions nicer
The slider was a bit wonky, having to stop at the right moment and release your finger to get the desired amount of questions, so I added a custom stepper component which has increment and decrement buttons and also has a TextInput which pops up a number keypad when pressed to quickly add a specific number. User cannot enter anything other than numbers and if they attempt to enter values lower than the minValue it simply enters the minValue, same for the maxValue. Has basic JSDoc documentation on the params that can be passed.